+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (feat) pass `cellEdit` in Table to react-bootstrap-table-next
 - (feat) allow ref to inner table and hide pagination in Table
 - (fix) Table header now pushes menu to right when hiding search
+- (fix) FieldSelect height consistent with other fields / use scss variables
   
 ## v0.4.2 (2019-08-02)